[Update] Small package naming refactors
