+++ Mobile/tues_pairs/lib/screens/main/match.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:tues_pairs/services/auth.dart';
 import 'package:tues_pairs/screens/main/settings.dart';
 import 'package:tues_pairs/screens/main/chat.dart';
-import 'package:tues_pairs/widgets/user_display_widgets/user_list.dart';
+import 'package:tues_pairs/widgets/user_display/user_list.dart';
 import 'package:provider/provider.dart';
 import 'package:tues_pairs/services/database.dart';
 import 'package:tues_pairs/modules/user.dart';
-import 'package:tues_pairs/widgets/user_display_widgets/already_matched.dart';
+import 'package:tues_pairs/widgets/user_display/already_matched.dart';
 
 class Match extends StatefulWidget {
   @override
