+++ Mobile/tues_pairs/lib/widgets/form/confim_password_input_field.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:tues_pairs/modules/user.dart';
 import 'package:tues_pairs/shared/constants.dart';
-import 'package:tues_pairs/widgets/form_widgets/input_field.dart';
+import 'package:tues_pairs/widgets/form/input_field.dart';
 
-class PasswordInputField extends InputField {
+class ConfirmPasswordInputField extends InputField {
 
-  PasswordInputField({Function onChanged}) : super(onChanged: onChanged);
-
+  ConfirmPasswordInputField({Function onChanged}) : super(onChanged: onChanged);
 
   @override
   Widget build(BuildContext context) {
@@ -14,14 +12,14 @@ class PasswordInputField extends InputField {
       obscureText: true, // obscures text (like for a password)
       onChanged: onChanged,
       style: textInputColor,
-      validator: (value) => value.isEmpty ? 'Enter a password' : null,
+      validator: (value) => value.isEmpty ? 'Confirm password' : null,
       keyboardType: TextInputType.visiblePassword,
       decoration: textInputDecoration.copyWith(
         icon: Icon(
-          Icons.lock,
+          Icons.repeat,
           color: Colors.orange,
         ),
-        hintText: 'Enter a password',
+        hintText: 'Confirm password',
       ),
     );
   }
